[S12P31A206-150] feat: CICDresolver 통합 테스트 진행(gitlab&jenkins api를 통한 자료 생성, AI 호출을 통한 파일 수정, 적용 및 재배포, 결과 보고서 등록)

See merge request s12-final/S12P31A206!241

+++ backend/src/main/java/org/example/backend/domain/aireport/entity/AIDeploymentReport.java
@@ -23,7 +23,7 @@ public class AIDeploymentReport {
 
     private Long projectId;
 
-    private Long buildNumber;
+    private int buildNumber;
 
     private String title;
 
@@ -42,7 +42,7 @@ public class AIDeploymentReport {
 
     private LocalDateTime createdAt;
 
-    public static AIDeploymentReport fromAiReportResponse(Long projectId, Long buildNumber,String title, AIReportResponse response, String commitUrl, String mergeRequestUrl, ReportStatus status) {
+    public static AIDeploymentReport fromAiReportResponse(Long projectId, int buildNumber,String title, AIReportResponse response, String commitUrl, String mergeRequestUrl, ReportStatus status) {
         return AIDeploymentReport.builder()
                 .projectId(projectId)
                 .buildNumber(buildNumber)
